ollama_gradio_project/app/requirements.txt
gradio
requests
psycopg2-binary
gunicorn
gevent
python-dotenv
